add notification feature and clean up

+++ src/app/(server)/api/product/delete/[id]/route.ts
@@ -1,9 +1,42 @@
+import notificationModel from "@/mongodb/notifications.model";
 import userProduct from "@/mongodb/product.model";
+import userAllProducts from "@/mongodb/userAllProducts.model";
+import cloudinary from "cloudinary";
 import { NextRequest, NextResponse } from "next/server";
+
+cloudinary.v2.config({
+    cloud_name: "deqs6ry98",
+    api_key: process.env.CLOUDINARY_APIKEY as string,
+    api_secret: process.env.CLUDINARY_APISECRET as string,
+    secure: true
+})
 
 export async function GET(req: NextRequest) {
     try {
         const _id = req.nextUrl.pathname.split("/").slice(-1)[0];
+
+        const product = await userProduct.findById({ _id });
+
+        if (!product) {
+            return NextResponse.json({ error: "Invalid request" }, { status: 400 });
+        }
+
+        const imageUniqueId = product.productImgURL.split("/").slice(-1)[0].slice(0, -4);
+
+        await cloudinary.v2.uploader.destroy(imageUniqueId);
+
+        const newNotification = await notificationModel.create({
+            notificationType: "delete",
+            notificationMessage: `Product '${product.Product_Details.product_name}' has been successfully deleted from the Analysis - Recent Added Products.`
+        });
+
+        let userAllOperations = await userAllProducts.findOne({
+            userEmail: product.userEmail
+        });
+
+        userAllOperations.appNotifications.push(newNotification._id);
+
+        userAllOperations.save();
 
         const deleteProduct = await userProduct.findOneAndDelete({
             _id
@@ -13,7 +46,7 @@ export async function GET(req: NextRequest) {
             return NextResponse.json({ error: "Invalid request" }, { status: 400 });
         }
 
-        return NextResponse.json({ success: true, message: "Product deleted successfully" });
+        return NextResponse.json({ success: true, message: "Product deleted successfully", notification: newNotification });
 
 
     } catch (err) {
